+++ src/Connection/ConnectionInterface.php
@@ -220,7 +220,7 @@ interface ConnectionInterface
      *
      * @return mixed Result of callback function.
      *
-     * @psalm-param Closure(ConnectionInterface): mixed $closure
+     * @psalm-param Closure(ConnectionInterface):mixed|Closure(ConnectionInterface):void $closure
      */
     public function transaction(Closure $closure, string $isolationLevel = null): mixed;
 }