+++ src/Connection/ConnectionInterface.php
@@ -212,13 +212,15 @@ interface ConnectionInterface
     /**
      * Executes callback provided in a transaction.
      *
-     * @psalm-param Closure(ConnectionInterface): mixed $closure A valid PHP callback that performs the job. Accepts connection instance as parameter.
+     * @param Closure $closure A valid PHP callback that performs the job. Accepts connection instance as parameter.
      * @param string|null $isolationLevel The isolation level to use for this transaction.
      * {@see TransactionInterface::begin()} for details.
      *
      * @throws Throwable If there is any exception during query. In this case, the transaction will be rolled back.
      *
      * @return mixed Result of callback function.
+     *
+     * @psalm-param Closure(ConnectionInterface):mixed|Closure(ConnectionInterface):void $closure
      */
     public function transaction(Closure $closure, string $isolationLevel = null): mixed;
 }