+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.1.1 under development
 
 - New #617: Add debug collector for `yiisoft/yii-debug` (@xepozz)
+- Enh #617, #733: Add specific psalm annotation of `$closure` parameter in `ConnectionInterface::transaction()` 
+  method (@xepozz, @vjik)
 
 ## 1.1.0 July 24, 2023