+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## 1.1.1 under development
 
-- no changes in this release.
+- New #617: Add debug collector for `yiisoft/yii-debug` (@xepozz)
+- Enh #617, #733: Add specific psalm annotation of `$closure` parameter in `ConnectionInterface::transaction()` 
+  method (@xepozz, @vjik)
 
 ## 1.1.0 July 24, 2023
 
-- Enh #617: Add debug collector for yiisoft/yii-debug (@xepozz)
 - Chg #722: Remove legacy array syntax for typecast. Use `Param` instead (@terabytesoftw)
 - Chg #724: Typecast refactoring (@Tigrov)
 - Chg #728: Refactor `AbstractSchema::getColumnPhpType()` (@Tigrov)