+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 # Yii Database Change Log
 
-## 1.1.1 under development
+## 1.1.1 August 16, 2023
 
-- Bug #741: Fix `alterColumn()` method to accept `ColumnInterface::class` in argument `$type` (@terabytesoftw)
 - New #617: Add debug collector for `yiisoft/yii-debug` (@xepozz)
 - Enh #617, #733: Add specific psalm annotation of `$closure` parameter in `ConnectionInterface::transaction()` 
   method (@xepozz, @vjik)
+- Bug #741: Fix `alterColumn()` method to accept `ColumnInterface::class` in argument `$type` (@terabytesoftw)
 
 ## 1.1.0 July 24, 2023